add jpu deeplab model

+++ utils/generator.py
@@ -162,7 +162,7 @@ class DataGenerator(keras.utils.Sequence):
                         ], p=0.3)
         ], p=1)
 
-        composed = composition(image=img, mask=masks)
+        composed = composition(image=img.astype('uint8'), mask=masks)
         aug_img = composed['image']
         aug_masks = composed['mask']
